Move WP Consent API integration to separate file
